move maps to their own directory
