+++ src/components/maps/bio/BioMapInfo.tsx
