+++ src/components/maps/collection/CollectionMapInfo.tsx
