+++ src/components/maps/rarity/RarityMap.css
