+++ src/components/maps/rarity/RarityMap.tsx
